requirements.txt
gradio>=5.0.0
requests>=2.32.0
python-dotenv>=1.0.0